// src/app/modules/academics/student-information-system/model/student-course-selection-view.ts
/**
 * 
 */
export interface StudentCourseSelectionView { 
    /**
     * Gets or Sets CourseCategoryId
     */
    courseCategoryId?: number;
    /**
     * Gets or Sets CourseIds
     */
    courseIds?: Array<string>;
    /**
     * Gets or Sets SchoolAcademicSessionStudentIds
     */
    schoolAcademicSessionStudentIds?: Array<string>;
    /**
     * Gets or Sets SchoolBoardId
     */
    schoolBoardId?: number;
    /**
     * Gets or Sets ClassId
     */
    classId?: number;
}